Update 09.0 WIP 18

+++ SP1Framework/pellet.cpp
@@ -1,6 +1,6 @@
 #include "pellet.h"
 
-SPellet::SPellet(COORD * c, short direction, double velocity, bool friendly)
+SPellet::SPellet(COORD * c, int direction, double velocity, bool friendly)
 {
 	this->m_dTime = 0.0;
 	this->m_cLocation = *c;
